Add files via upload
Tells the amount of RLPs/RLKs in my selection with and without extracellular domain, then tells the amount of proteins that have a certain extracellular domain.
Something to take into account: Of the RLPs without an extracellular domain, not all are RLPs. Some are likely proteins in the membrane with a different function.